Fix ls support
Safe doesn't use HTTP verb "LIST", but instead uses GET and adds
list=true as a query parameter.  The LIST method subroute wasn't getting
triggered, so added clause to check for query param in the GET handler
and redirect to the LIST handler.

Also some cleanup of logging and a more appropriate name for the mount
structure.